speed improvements to file conversion, some refactoring/fixes
regexes and operator.itemgetter() are now used to parse and access the
genotypes for the selected samples. from profiling this has been found
to be faster than numpy index vectors, and the required casting, also
faster than list comprehensions. lines are now written out in blocks,
to take advantage of HDD dwell and SSD paging that emulates HDD
behavior.