improvements to encoding and multi-allelic merges
genotype coding function can now specified in a separate python file
via the —codingFunctionClassFile, containing a GenoCoder class with a
coding_function(current_value, reference_allele, ancestral_allele) that
returns an array of “0” and “1” for selscan. The method used to combine
coded output for each allele of a multiallelic site is a bitwise
operator, now specified via the hidden —multiAllelicMergeFunction
argument. It accepts OR, AND, XOR. It is currently set to a default of
OR, though I believe it should perhaps be AND.