when encoding tpeds from VCF input, skip loci that incorrectly encode unknown genotypes
Sometimes uncalled genotypes are represented per-sample as ‘.’, when
they should be ‘.|.’ for phased haploid genomes. Per the VCF spec: “If
a call cannot be made for a sample at a given locus, ‘.’ should be
specified for each missing allele in the GT field (for example ‘./.’
for a diploid genotype and ‘.’ for haploid genotype).” This change
skips such lines since we can’t know anything about the loci anyway,
and continues encoding the other lines in the file.